backend/requirements.txt
asgiref==3.8.1
Django==5.0.8
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
phonenumbers==8.13.42
psycopg==3.2.1
psycopg-binary==3.2.1
sqlparse==0.5.1
typing_extensions==4.12.2
